+++ bulletin_board/js-client/src/voter/voter.js
@@ -78,6 +78,7 @@ export class Voter {
       encryptedData,
       encryptedDataHash,
       auditableData,
+      plainVote,
     };
   }