+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+
+- The voter `encrypt` method now returns the `plainVote` so it will be included when the user audits the ballot.
+
 ## [0.18.0] - 2021-04-16
 
 ## Added